requirements.txt
python-docx>=1.1.0
python-dotenv>=1.0.0
anthropic>=0.20.0
